feat: all stderr outputs normalized; idempotent downloads to avoid redownloads; tools versioning

+++ lib/downloaders/gene.sh
@@ -4,29 +4,19 @@ downloaders_gene::run_gene_download() {
 
     downloaders_common::prepare_download_environment
 
+    local input tmp
     if [[ "$GENE_ID_USER" == true ]]; then
-
-        local tmp
         tmp=$(mktemp)
-
-        echo "$GENE_ID" |
-            tr ',' '\n' > "$tmp"
-
+        echo "$GENE_ID" | tr ',' '\n' > "$tmp"
         input="$tmp"
-
     else
-
         input="$GENE_FILE"
-
     fi
 
-    downloaders_ncbi_download::download_genes_batches \
-        "$input" \
-        "$OUTDIR" \
-        "$PARALLEL_JOBS"
+    local rc=0
+    downloaders_ncbi_download::download_genes_batches "$input" "$OUTDIR" "$PARALLEL_JOBS" || rc=$?
 
-    [[ "$GENE_ID_USER" == true ]] &&
-        rm -f "$tmp"
-
+    [[ "$GENE_ID_USER" == true ]] && rm -f "$tmp"
     downloaders_common::cleanup_temp
+    return "$rc"
 }
